add autherizartion

+++ app/Http/Controllers/NoteController.php
@@ -51,6 +51,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id !== auth()->id()) {
+            abort(403);
+        }
         return view("note.show", compact("note"));
     }
 
@@ -59,6 +62,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id !== auth()->id()) {
+            abort(403);
+        }
         return view("note.edit", compact("note"));
     }
 
@@ -67,6 +73,9 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id !== auth()->id()) {
+            abort(403);
+        }
         $data = $request->validate([
             "note" => ["required", "string"],
         ]);
@@ -84,6 +93,10 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id !== auth()->id()) {
+            abort(403);
+        }
+
         $note->delete();
 
         return to_route("note.index")->with(
